requirements.txt
numpy==1.26.2
openai==0.28.1
PyQt6==6.6.0
regex==2023.10.3
tqdm==4.66.1
python-dotenv
